requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
dataclasses-json==0.6.6
distro==1.9.0
exceptiongroup==1.2.1
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jsonpatch==1.33
jsonpointer==2.4
langchain==0.2.0
langchain-core==0.2.1
langchain-text-splitters==0.2.0
langsmith==0.1.60
marshmallow==3.21.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.30.1
orjson==3.10.3
packaging==23.2
pydantic==2.7.1
pydantic_core==2.18.2
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.32.2
sniffio==1.3.1
SQLAlchemy==2.0.30
tenacity==8.3.0
tqdm==4.66.4
typing-inspect==0.9.0
typing_extensions==4.11.0
urllib3==2.2.1
yarl==1.9.4
